requirements.txt
anthropic
pandas
requests
openpyxl
streamlit
python-dotenv
urllib3
